Moved SftpDownloader to core.
