+++ core/pom.xml
@@ -178,6 +178,13 @@
             <version>2.11.0</version>
         </dependency>
 
+        <dependency>
+            <groupId>com.jcraft</groupId>
+            <artifactId>jsch</artifactId>
+            <version>0.1.55</version>
+        </dependency>
+
+
     </dependencies>
 
 </project>
