+++ tasks/pom.xml
@@ -49,11 +49,5 @@
             <artifactId>opencsv</artifactId>
             <version>4.5</version>
         </dependency>
-
-        <dependency>
-            <groupId>com.jcraft</groupId>
-            <artifactId>jsch</artifactId>
-            <version>0.1.55</version>
-        </dependency>
     </dependencies>
 </project>
