+++ tasks/src/main/java/com/lantanagroup/link/tasks/config/ParklandInventoryImportConfig.java
@@ -1,6 +1,6 @@
 package com.lantanagroup.link.tasks.config;
 
-import com.lantanagroup.link.tasks.config.SftpDownloaderConfig;
+import com.lantanagroup.link.config.SftpDownloaderConfig;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
